requirements.txt
streamlit>=1.0
tensorflow>=2.8
numpy
pillow